Add files via upload
The unittest_template.py is added to this repository as a reference and easy starting point for creating unit tests in Python. It contains all the basics and can be run on the also added fibonacci.py file to see how it works in action. The template can be used as it is as a starting point for any basic unit testing, just modifying the test cases to match the intended target to be tested.